py4seo/Код с занятий/lesson6/requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
Django==2.2.7
fake-useragent==0.1.11
idna==2.8
lxml==4.4.1
multidict==4.5.2
parse==1.12.1
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
python-slugify==4.0.0
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.3
tqdm==4.37.0
urllib3==1.25.6
w3lib==1.21.0
websockets==8.1
yarl==1.3.0
